+++ backend-app/src/main/java/mouse/univ/backendapp/controller/ShopController.java
@@ -2,6 +2,7 @@ package mouse.univ.backendapp.controller;
 
 import lombok.RequiredArgsConstructor;
 import mouse.univ.backendapp.api.ApiResponse;
+import mouse.univ.backendapp.auth.controller.UD;
 import mouse.univ.backendapp.dto.shop.ShopCreateDTO;
 import mouse.univ.backendapp.dto.shop.ShopResponseDTO;
 import mouse.univ.backendapp.dto.shop.ShopUpdateDTO;
@@ -20,35 +21,37 @@ public class ShopController {
     private final ShopService shopService;
 
     @PostMapping
-    public ResponseEntity<ApiResponse<ShopResponseDTO>> postProduct(@RequestBody ShopCreateDTO shopCreateDTO) {
+    public ResponseEntity<ApiResponse<ShopResponseDTO>> postShop(@RequestBody ShopCreateDTO shopCreateDTO, @RequestAttribute("user") UserDetails userDetails) {
+        UD.validateAdmin(userDetails);
         ShopResponseDTO result = shopService.createShop(shopCreateDTO);
         ApiResponse<ShopResponseDTO> apiResponse = ApiResponse.ok(result);
         return ResponseEntity.status(HttpStatus.CREATED).body(apiResponse);
     }
 
     @PutMapping
-    public ResponseEntity<ApiResponse<ShopResponseDTO>> updateProduct(@RequestBody ShopUpdateDTO shopUpdateDTO) {
+    public ResponseEntity<ApiResponse<ShopResponseDTO>> updateShop(@RequestBody ShopUpdateDTO shopUpdateDTO, @RequestAttribute("user") UserDetails userDetails) {
+        UD.validateAdmin(userDetails);
         ShopResponseDTO result = shopService.updateShop(shopUpdateDTO);
         ApiResponse<ShopResponseDTO> apiResponse = ApiResponse.ok(result);
         return ResponseEntity.status(HttpStatus.OK).body(apiResponse);
     }
     @GetMapping
-    public ResponseEntity<ApiResponse<List<ShopResponseDTO>>> getAllProducts() {
+    public ResponseEntity<ApiResponse<List<ShopResponseDTO>>> getAllShops() {
         List<ShopResponseDTO> result = shopService.getAllShops();
         ApiResponse<List<ShopResponseDTO>> apiResponse = ApiResponse.ok(result);
         return ResponseEntity.status(HttpStatus.OK).body(apiResponse);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ApiResponse<ShopResponseDTO>> getAllProducts(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse<ShopResponseDTO>> getAllShops(@PathVariable Long id) {
         ShopResponseDTO result = shopService.findShopById(id);
         ApiResponse<ShopResponseDTO> apiResponse = ApiResponse.ok(result);
         return ResponseEntity.status(HttpStatus.OK).body(apiResponse);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<ApiResponse<Object>> deleteProductById(@PathVariable Long id) {
-        UserDetails userDetails = UserDetails.asAdmin();
+    public ResponseEntity<ApiResponse<Object>> deleteShopById(@PathVariable Long id, @RequestAttribute("user") UserDetails userDetails) {
+        UD.validateAdmin(userDetails);
         shopService.deleteShopById(id, userDetails);
         ApiResponse<Object> apiResponse = ApiResponse.ok(null);
         return ResponseEntity.status(HttpStatus.OK).body(apiResponse);
